Change name of CustomSpinner to ScrolledToTopSpinner (#1955)
