+++ collect_app/src/main/res/layout/spinner_layout.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<org.odk.collect.android.views.CustomSpinner xmlns:android="http://schemas.android.com/apk/res/android"
+<org.odk.collect.android.views.ScrolledToTopSpinner xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/spinner"
     style="@style/Base.Widget.AppCompat.Spinner.Underlined"
     android:layout_width="match_parent"
